Changes to PlayerController and SpawnPoint PlayerPref implemented
Really confused on how arrays work now.

If a spawnpoint player pref hasn't been made, it will default to 0.

When the scene begins, it will bring the player to the corresponding spawn point. eg. spawnPoint[0]

However, the array in the unity editor starts from 1. So spawnPoint[0] should set in spawnPoint item 1, but it doesn't work like that apparently